Moved most difficulty code into its own package. Also added lists the entity config data.
