--- src/main/java/com/vortexel/dangerzone/common/ModifierType.java
@@ -1,18 +1,0 @@
-package com.vortexel.dangerzone.common;
-
-public enum ModifierType {
-    MAX_HEALTH,
-    REGENERATION,
-
-    ATTACK_DAMAGE,
-    ATTACK_SPEED,
-
-    MOVE_SPEED,
-    FLY_SPEED,
-
-    ARMOR,
-    ARMOR_TOUGHNESS,
-
-    WITHER,
-    EXPLOSION_RADIUS,
-}
